requirements.txt
litellm
python-dotenv
pandas